.faaspot/requirements.txt
slackclient==1.1.0
python-box==3.1.1
tmdbv3api==0.5